@@ -4,7 +4,7 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "node scripts/free-port.js && node scripts/clean-next.js && next dev",
+    "dev": "node scripts/free-port.js && node scripts/clean-next.js && next dev -p 3001",
     "dev:3001": "node scripts/free-port.js && node scripts/clean-next.js && next dev -p 3001",
     "build": "node scripts/clean-next.js && next build",
     "start": "next start",
@@ -38,7 +38,7 @@
     "@fullcalendar/react": "^6.1.17",
     "@fullcalendar/timegrid": "^6.1.17",
     "@heroicons/react": "^2.2.0",
-    "@prisma/client": "^6.7.0",
+    "@prisma/client": "^6.11.0",
     "@radix-ui/react-alert-dialog": "^1.1.11",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-checkbox": "^1.2.3",
@@ -55,7 +55,6 @@
     "@radix-ui/react-tabs": "^1.1.11",
     "@radix-ui/react-tooltip": "^1.2.7",
     "@stripe/stripe-js": "^7.2.0",
-    "@types/bcryptjs": "^2.4.6",
     "@types/node-cron": "^3.0.11",
     "@types/winston": "^2.4.4",
     "@upstash/redis": "^1.35.0",
@@ -69,10 +68,11 @@
     "dotenv": "^16.5.0",
     "lucide-react": "^0.503.0",
     "multer": "^2.0.1",
-    "next": "14.2.0",
+    "next": "^15.3.4",
     "next-auth": "^4.24.11",
     "next-themes": "^0.4.6",
     "node-cron": "^4.1.1",
+    "node-fetch": "^3.3.2",
     "prom-client": "^15.1.3",
     "puppeteer": "^24.8.1",
     "react": "^18.2.0",
@@ -82,7 +82,7 @@
     "react-dropzone": "^14.3.8",
     "react-icons": "^5.5.0",
     "recharts": "^2.15.3",
-    "resend": "^4.5.1",
+    "resend": "^4.6.0",
     "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1",
     "sonner": "^2.0.5",
@@ -99,6 +99,7 @@
     "@eslint/eslintrc": "^3",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^14.3.1",
+    "@types/bcryptjs": "^2.4.6",
     "@types/jest": "^29.5.14",
     "@types/node": "^20",
     "@types/react": "19.1.8",
@@ -116,8 +117,8 @@
     "pngjs": "^7.0.0",
     "postcss": "^8.5.3",
     "postcss-nesting": "^13.0.1",
-    "prisma": "^6.7.0",
-    "tailwindcss": "^3.4.1",
+    "prisma": "^6.11.0",
+    "tailwindcss": "^3.4.3",
     "typescript": "^5"
   },
   "prisma": {
